Refactor: Update import paths to remove 'src.' prefix for consistency
